Moved BetterBadlandsRender.java to its own client package and renamed it, because now it looks really cool and handsome.
